feat: unopinionated App::view rendering
The application loop behavior from the original rustea library would
take the String result from calling App::view and print it to the screen
while also clearing previously printed lines. This meant the application
only needed to focus on how to construct the appropriate multiline
String to render on the screen, and the application event loop would
gracefully handle the rest.

While this did provide an easy onboarding to the library to get up and
running, it also introduced some weight to the amount of data being
streamed out to the terminal to perform clearing and re-printing of
lines which may not have changed between application state updates.  For
applications which fill the screen but have minimal changes to the view
on updates, this can degrade performance.  For this reason, this default
behavior was removed, and it is now expected the application will
provide its own logic for how to print to the screen.  The App::view
signature now has no return type and it is still not allowed to mutate
the state of the Model.

This new design allows an application to finely tune cursor movements
and printing logic to only target portions of the view that need
updated.  The previous behavior of just printing new line separated
Strings is still useful for simple scenarios though and that logic was
extracted into `view_helper/string_printer.rs`.  An example of it in use
can be found in `examples/input.rs`.